Organized the FP tests.
